+++ internal/models/withdrawals.go
@@ -1,8 +1,8 @@
 package models
 
 type Withdrawals struct {
-	ID          string `json:"order"` // использовать как primary key
-	UserId      string `json:"-"`
-	Sum         uint   `json:"sum"`
-	ProcessedAt string `json:"processed_at"`
+	ID          string  `json:"order"` // использовать как primary key
+	UserId      string  `json:"-"`
+	Sum         float64 `json:"sum"`
+	ProcessedAt string  `json:"processed_at"`
 }
